Implement dom-display.js, dom-actions.js, tests.test.js, style.css, media-query.css, and template.html
There weren't many commits for this project as I procrastinated and
considered changing the framework many times. As you may see in
index.js, I ended up using my initial plan. Jest testing with babel was
used for testing coordinates but, due to conflict with the webpack
bundler and babel's ESM testing, much of the code was too simple to
necessitate testing or made use of DOM interaction in a way that
couldn't be further depricated. However, testing was used for
displayShip's handling of offsets.

Now knowing that I could have offset the player boards, as I chose to do
with the radar and sonar grids, I can see that this is the preferable
method for maintaining the ability to use mix-blend-mode in CSS.
However, dom-actions and dom-display don't leave much to be desired.
They utilize factory and class encapsulation effectively to be accessed
for computation and DOM manipulation, including comprehensive promises
and AJAX, to ensure that the player can run multiple games with
relative ease.

Implement dom-actions
  -  Utilize startGame -> displayBoard -> quickDeploy || placeFleet ->
     gameLoop -> addSquareListeners (with checkHit as a callback)
  -  The aforemention flow of execution utilizes dom-display.js
     functions to append relevant information to the DOM

Implement dom-display
  -  Utilize eventListeners to control most callbacks, resolves, and returns
     to dom-actions